% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ectd.R
\name{ectd}
\alias{ectd}
\title{Compute the Euclidian Commute-Time Distances (ECTD) in an undirected graph.}
\usage{
ectd(o, similarity = function(x) 1, inf.replace = TRUE)
}
\arguments{
\item{o}{Output of \code{nng}.}

\item{similarity}{function distance --> similarity.}

\item{inf.replace}{Replace Inf values by large finite number ?}
}
\value{
A distances matrix (n x n)
}
\description{
Assuming similarity function doesn't depend on x, and undirected graph.
}
