% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nng.R
\name{nng}
\alias{nng}
\title{Compute the nearest-neighbors graph + pairwise distances.}
\usage{
nng(data, k = ceil(sqrt(nrow(data))), mutual = TRUE)
}
\arguments{
\item{data}{data.frame or matrix.}

\item{k}{Number of neighbors at each point.}

\item{mutual}{Whether or not to build mutual kNN graph.}
}
\value{
A list with $graph and $distances.
}
\description{
This is just a wrapper around C++/Rcpp function.
}
