\name{nngd-package}
\alias{nngd-package}
\alias{nngd}
\docType{package}

\title{
	\packageTitle{nngd}
}

\description{
	\packageDescription{nngd}
}

\details{
  TODO. See functions \code{nng()} and \code{ectd()}.
}

\author{
	\packageAuthor{nngd}

	Maintainer: \packageMaintainer{nngd}
}

%\references{
%	TODO: Literature or other references for background information
%}

%\examples{
%	TODO: simple examples of the most important functions
%}
